refactor(py): replace zarr-python type helpers with self-contained store types
Add _store_types.py defining ngff-zarr's own StoreLike (str | Path |
os.PathLike | MutableMapping) and repoint every importer of the deleted
_zarr_types module to it. The config default cache_store becomes the
cache directory path itself, targeted directly by the zarrista-backed
cache writer, so config no longer imports zarr. The _zarr_kwargs and
_zarr_open_array compat modules are deleted; their definitions are
inlined next to the remaining legacy zarr-python fallback branches in
to_multiscales.py and to_ngff_zarr.py, which are slated for removal
with those branches.
