+++ py/ngff_zarr/config.py
@@ -5,26 +5,13 @@ from pathlib import Path
 
 import dask.config
 from platformdirs import user_cache_dir
-from zarr.storage import StoreLike
 
-from ._zarr_kwargs import zarr_kwargs
+from ._store_types import StoreLike
 
 if dask.config.get("temporary-directory") is not None:
     _store_dir = dask.config.get("temporary-directory")
 else:
     _store_dir = Path(user_cache_dir("ngff-zarr"))
-
-
-def default_store_factory():
-    try:
-        from zarr.storage import DirectoryStore
-
-        return DirectoryStore(_store_dir, **zarr_kwargs)
-    except ImportError:
-        from zarr.storage import LocalStore
-
-        return LocalStore(_store_dir)
-
 
 try:
     import psutil
@@ -39,7 +26,9 @@ class NgffZarrConfig:
     # Rough memory target in bytes
     memory_target: int = default_memory_target
     task_target: int = 50000
-    cache_store: StoreLike = field(default_factory=default_store_factory)
+    # The cache directory path; large-image cache arrays are written into it
+    # through zarrista.
+    cache_store: StoreLike = field(default_factory=lambda: _store_dir)
 
     # HCS cache settings
     hcs_image_cache_size: int = 100  # Max number of images to cache per well
